Change folder structure
Under the new folder structure, all source files that are used by both
the client and the server should go in the src/ folder.  Sources files
only used by one of the two should go in src/client/ or src/server/.
The include/ folder, which contains the header files, mirrors this
structure.

The makefile was updated to reflect the changes.  It is no longer
necessary to put every dependency into the makefile separately.  This
was the main motivation for these changes in the folder structure.