Make client receiving and sending synchronous
This means we send and receive in the same thread, no longer in two
different threads.  The reason for this is that in the commands `get`
and `put`, we need to know what response belongs to which command.
Since TCP doesn't give us any timing guarantees, we still can't be
sure that we have it right, it's just more likely this way.  We
probably can't do any better because of the way the GRASS protocol is.

This change is super ugly and I wish we didn't have to do it.  We now
have to use timeouts to detect when the end of an incoming message has
come.  Maybe there is a better solution, but I don't see it.